Implemented PR feedback
