+++ packages/core/src/api/blockManipulation/commands/moveBlocks/moveBlocks.ts
@@ -26,12 +26,15 @@ type BlockSelectionData = (
   anchorBlockId: string;
 };
 
-// `getBlockSelectionData` and `updateBlockSelectionFromData` are used to save
-// and restore the selection within a block, when the block is moved. This is
-// done by first saving the offsets of the anchor and head from the before
-// positions of their surrounding blocks, as well as the IDs of those blocks. We
-// can then recreate the selection by finding the blocks with those IDs, getting
-// their before positions, and adding the offsets to those positions.
+/**
+ * `getBlockSelectionData` and `updateBlockSelectionFromData` are used to save
+ * and restore the selection within a block, when the block is moved. This is
+ * done by first saving the offsets of the anchor and head from the before
+ * positions of their surrounding blocks, as well as the IDs of those blocks. We
+ * can then recreate the selection by finding the blocks with those IDs, getting
+ * their before positions, and adding the offsets to those positions.
+ * @param editor The BlockNote editor instance to get the selection data from.
+ */
 function getBlockSelectionData(
   editor: BlockNoteEditor<any, any, any>
 ): BlockSelectionData {
@@ -67,12 +70,17 @@ function getBlockSelectionData(
   }
 }
 
-// `getBlockSelectionData` and `updateBlockSelectionFromData` are used to save
-// and restore the selection within a block, when the block is moved. This is
-// done by first saving the offsets of the anchor and head from the before
-// positions of their surrounding blocks, as well as the IDs of those blocks. We
-// can then recreate the selection by finding the blocks with those IDs, getting
-// their before positions, and adding the offsets to those positions.
+/**
+ * `getBlockSelectionData` and `updateBlockSelectionFromData` are used to save
+ * and restore the selection within a block, when the block is moved. This is
+ * done by first saving the offsets of the anchor and head from the before
+ * positions of their surrounding blocks, as well as the IDs of those blocks. We
+ * can then recreate the selection by finding the blocks with those IDs, getting
+ * their before positions, and adding the offsets to those positions.
+ * @param editor The BlockNote editor instance to update the selection in.
+ * @param data The selection data to update the selection with (generated by
+ * `getBlockSelectionData`).
+ */
 function updateBlockSelectionFromData(
   editor: BlockNoteEditor<any, any, any>,
   data: BlockSelectionData
@@ -112,9 +120,12 @@ function updateBlockSelectionFromData(
   );
 }
 
-// Replaces any `columnList` blocks with the children of their columns. This is
-// done here instead of in `getSelection` as we still need to remove the entire
-// `columnList` node but only insert the `blockContainer` nodes inside it.
+/**
+ * Replaces any `columnList` blocks with the children of their columns. This is
+ * done here instead of in `getSelection` as we still need to remove the entire
+ * `columnList` node but only insert the `blockContainer` nodes inside it.
+ * @param blocks The blocks to flatten.
+ */
 function flattenColumns(
   blocks: Block<any, any, any>[]
 ): Block<any, any, any>[] {
@@ -134,9 +145,16 @@ function flattenColumns(
     .flat();
 }
 
-// Removes the selected blocks from the editor, then inserts them before/after a
-// reference block. Also updates the selection to match the original selection
-// using `getBlockSelectionData` and `updateBlockSelectionFromData`.
+/**
+ * Removes the selected blocks from the editor, then inserts them before/after a
+ * reference block. Also updates the selection to match the original selection
+ * using `getBlockSelectionData` and `updateBlockSelectionFromData`.
+ * @param editor The BlockNote editor instance to move the blocks in.
+ * @param referenceBlock The reference block to insert the selected blocks
+ * before/after.
+ * @param placement Whether to insert the selected blocks before or after the
+ * reference block.
+ */
 export function moveSelectedBlocksAndSelection(
   editor: BlockNoteEditor<any, any, any>,
   referenceBlock: BlockIdentifier,
@@ -155,9 +173,8 @@ export function moveSelectedBlocksAndSelection(
 
 // Checks if a block is in a valid place after being moved. This check is
 // primitive at the moment and only returns false if the block's parent is a
-// `columnList` block. This is because when moving blocks, columns are flattened
-// into regular blocks and therefore can only be nested inside `column` or
-// other regular blocks.
+// `columnList` block. This is because regular blocks cannot be direct children
+// of `columnList` blocks.
 function checkPlacementIsValid(parentBlock?: Block<any, any, any>): boolean {
   return !parentBlock || parentBlock.type !== "columnList";
 }
@@ -170,7 +187,7 @@ function checkPlacementIsValid(parentBlock?: Block<any, any, any>): boolean {
 // 3. If the block has no previous sibling, but is nested, the placement is
 // before its parent.
 // If the placement is invalid, the function is called recursively until a valid
-// placement is found. Returns undefined if no valid placement is found or if
+// placement is found. Returns undefined if no valid placement is found, meaning
 // the block is already at the top of the document.
 function getMoveUpPlacement(
   editor: BlockNoteEditor<any, any, any>,
@@ -195,6 +212,7 @@ function getMoveUpPlacement(
     placement = "before";
   }
 
+  // Case when the block is already at the top of the document.
   if (!referenceBlock || !placement) {
     return undefined;
   }
@@ -221,7 +239,7 @@ function getMoveUpPlacement(
 // 3. If the block has no next sibling, but is nested, the placement is
 // after its parent.
 // If the placement is invalid, the function is called recursively until a valid
-// placement is found. Returns undefined if no valid placement is found or if
+// placement is found. Returns undefined if no valid placement is found, meaning
 // the block is already at the bottom of the document.
 function getMoveDownPlacement(
   editor: BlockNoteEditor<any, any, any>,
@@ -246,6 +264,7 @@ function getMoveDownPlacement(
     placement = "after";
   }
 
+  // Case when the block is already at the bottom of the document.
   if (!referenceBlock || !placement) {
     return undefined;
   }
